Check Similarity Program
This program created to check the similarities between the 2 databases. The input file is a .txt file, which contains the data of the musical institutions. 
The program will search the catalog numbers in  2 databases, and when the catalog numbers are matched will compare the names of composers linked to them. The similarity is based on the Levenshtein Measure of distance.
The input data format is the following (URI)(separator ",,,")(catalogue_number)(separator ",,,")(composer_name)(new line)
The threshold value should be assigned in the prog.
The output is written into a new .txt file with measures of the similarity for each entry.